refactor(provider): convert provider details screen to bottom sheet

+++ data/downloads/src/main/kotlin/com/flixclusive/data/downloads/model/DownloadState.kt
@@ -6,7 +6,7 @@ import java.io.File
 data class DownloadState(
     val downloadId: String,
     val status: DownloadStatus = DownloadStatus.IDLE,
-    val progress: Int = 0,
+    val progress: Float = 0f,
     val bytesDownloaded: Long = 0L,
     val totalBytes: Long = 0L,
     val error: UiText? = null,
